+++ src/cql/InterventionLogic_Common.json
@@ -298,164 +298,6 @@
                               "pattern" : {
                                  "type" : "Lower",
                                  "operand" : {
-                                    "name" : "ToString",
-                                    "libraryName" : "FHIRHelpers",
-                                    "type" : "FunctionRef",
-                                    "operand" : [ {
-                                       "path" : "name",
-                                       "type" : "Property",
-                                       "source" : {
-                                          "name" : "currentQuestionnaire",
-                                          "type" : "OperandRef"
-                                       }
-                                    } ]
-                                 }
-                              },
-                              "string" : {
-                                 "type" : "Lower",
-                                 "operand" : {
-                                    "path" : "value",
-                                    "type" : "Property",
-                                    "source" : {
-                                       "path" : "questionnaire",
-                                       "scope" : "Q",
-                                       "type" : "Property"
-                                    }
-                                 }
-                              }
-                           }, {
-                              "type" : "Negate",
-                              "operand" : {
-                                 "valueType" : "{urn:hl7-org:elm-types:r1}Integer",
-                                 "value" : "1",
-                                 "type" : "Literal"
-                              }
-                           } ]
-                        }
-                     }, {
-                        "type" : "Not",
-                        "operand" : {
-                           "type" : "Equal",
-                           "operand" : [ {
-                              "type" : "PositionOf",
-                              "pattern" : {
-                                 "type" : "Lower",
-                                 "operand" : {
-                                    "name" : "ToString",
-                                    "libraryName" : "FHIRHelpers",
-                                    "type" : "FunctionRef",
-                                    "operand" : [ {
-                                       "path" : "id",
-                                       "type" : "Property",
-                                       "source" : {
-                                          "name" : "currentQuestionnaire",
-                                          "type" : "OperandRef"
-                                       }
-                                    } ]
-                                 }
-                              },
-                              "string" : {
-                                 "type" : "Lower",
-                                 "operand" : {
-                                    "path" : "value",
-                                    "type" : "Property",
-                                    "source" : {
-                                       "path" : "questionnaire",
-                                       "scope" : "Q",
-                                       "type" : "Property"
-                                    }
-                                 }
-                              }
-                           }, {
-                              "type" : "Negate",
-                              "operand" : {
-                                 "valueType" : "{urn:hl7-org:elm-types:r1}Integer",
-                                 "value" : "1",
-                                 "type" : "Literal"
-                              }
-                           } ]
-                        }
-                     } ]
-                  }, {
-                     "type" : "Equal",
-                     "operand" : [ {
-                        "path" : "value",
-                        "type" : "Property",
-                        "source" : {
-                           "path" : "status",
-                           "scope" : "Q",
-                           "type" : "Property"
-                        }
-                     }, {
-                        "valueType" : "{urn:hl7-org:elm-types:r1}String",
-                        "value" : "completed",
-                        "type" : "Literal"
-                     } ]
-                  } ]
-               },
-               "sort" : {
-                  "by" : [ {
-                     "direction" : "desc",
-                     "path" : "authored",
-                     "type" : "ByColumn"
-                  }, {
-                     "direction" : "desc",
-                     "type" : "ByExpression",
-                     "expression" : {
-                        "path" : "lastUpdated",
-                        "type" : "Property",
-                        "source" : {
-                           "name" : "meta",
-                           "type" : "IdentifierRef"
-                        }
-                     }
-                  } ]
-               }
-            },
-            "operand" : [ {
-               "name" : "responses",
-               "operandTypeSpecifier" : {
-                  "type" : "ListTypeSpecifier",
-                  "elementType" : {
-                     "name" : "{http://hl7.org/fhir}QuestionnaireResponse",
-                     "type" : "NamedTypeSpecifier"
-                  }
-               }
-            }, {
-               "name" : "currentQuestionnaire",
-               "operandTypeSpecifier" : {
-                  "name" : "{http://hl7.org/fhir}Questionnaire",
-                  "type" : "NamedTypeSpecifier"
-               }
-            } ]
-         }, {
-            "name" : "MatchedReponsesByQuestionnaire",
-            "context" : "Patient",
-            "accessLevel" : "Public",
-            "type" : "FunctionDef",
-            "expression" : {
-               "type" : "Query",
-               "source" : [ {
-                  "alias" : "Q",
-                  "expression" : {
-                     "name" : "responses",
-                     "type" : "OperandRef"
-                  }
-               } ],
-               "relationship" : [ ],
-               "where" : {
-                  "type" : "And",
-                  "operand" : [ {
-                     "type" : "Or",
-                     "operand" : [ {
-                        "type" : "Not",
-                        "operand" : {
-                           "type" : "Equal",
-                           "operand" : [ {
-                              "type" : "PositionOf",
-                              "pattern" : {
-                                 "type" : "Lower",
-                                 "operand" : {
                                     "name" : "QuestionnaireID",
                                     "type" : "OperandRef"
                                  }
