+++ src/cql/source/GAD7_InterventionLogicLibrary.cql
@@ -38,6 +38,9 @@ define CurrentQuestionnaire:
 define CurrentQuestionnaireName:
   QuestionnaireName
 
+define CurrentQuestionnaireID:
+  QuestionnaireID
+
 define QuestionnaireItems:
   CurrentQuestionnaire Q
   return Q.item
@@ -48,11 +51,8 @@ define sortedResponses:
   sort by authored desc
 // Load Questionnaire responses
 define QuestionnaireResponses:
-  if QuestionnaireItems is not null then 
-    LogicHelpers.MatchedReponsesByQuestionnaire([QuestionnaireResponse], CurrentQuestionnaire)
-  else
-    LogicHelpers.MatchedReponsesByQuestionnaire([QuestionnaireResponse], QuestionnaireID, QuestionnaireName)
-
+  LogicHelpers.MatchedReponsesByQuestionnaire([QuestionnaireResponse], QuestionnaireID, QuestionnaireName)
+ 
 define ResponsesSummary:
   (QuestionnaireResponses) I
   let
