+++ src/cql/source/PHQ9_InterventionLogicLibrary.cql
@@ -45,16 +45,16 @@ define CurrentQuestionnaire:
 define CurrentQuestionnaireName:
   QuestionnaireName
 
+define CurrentQuestionnaireID:
+  QuestionnaireID
+
 define QuestionnaireItems:
   CurrentQuestionnaire Q
   return Q.item
 
 // Load Questionnaire responses
 define QuestionnaireResponses:
-  if QuestionnaireItems is not null then 
-    LogicHelpers.MatchedReponsesByQuestionnaire([QuestionnaireResponse], CurrentQuestionnaire)
-  else
-    LogicHelpers.MatchedReponsesByQuestionnaire([QuestionnaireResponse], QuestionnaireID, QuestionnaireName)
+  LogicHelpers.MatchedReponsesByQuestionnaire([QuestionnaireResponse], QuestionnaireID, QuestionnaireName)
 
 define ResponsesSummary:
   (QuestionnaireResponses) I
